requirements.txt
fastapi==0.120.4
uvicorn==0.38.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic[email]==2.12.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.1.1
httpx==0.26.0
aiofiles==23.2.1

# Email support
email-validator==2.3.0
resend>=2.0.0
